refactor(bigint): separate operations
- moved separate operations into their own separate files, making the bigint folder structure make more sense
