+++ src/main.cpp
@@ -1,6 +1,5 @@
 #include "lexer/lexer.hpp"
 #include "core/core.hpp"
-#include <iostream>
 #include "bigint/bigint.hpp"
 
 int main(){
